# Add minimum length optimization to RegexCompiler