--- a/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexTree.cs
+++ b/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexTree.cs
@@ -19,8 +19,9 @@ namespace System.Text.RegularExpressions
         public readonly Hashtable CapNames;
         public readonly string[] CapsList;
         public readonly RegexOptions Options;
+        public readonly int MinRequiredLength;
 
-        internal RegexTree(RegexNode root, Hashtable caps, int[] capNumList, int capTop, Hashtable capNames, string[] capsList, RegexOptions options)
+        internal RegexTree(RegexNode root, Hashtable caps, int[] capNumList, int capTop, Hashtable capNames, string[] capsList, RegexOptions options, int minRequiredLength)
         {
             Root = root;
             Caps = caps;
@@ -29,6 +30,7 @@ namespace System.Text.RegularExpressions
             CapNames = capNames;
             CapsList = capsList;
             Options = options;
+            MinRequiredLength = minRequiredLength;
         }
 
 #if DEBUG