We now do a cursory examination of the RegexNode tree in order to determine the smallest length string that could possibly match.  For example, the regex "(?:123|4567)8" has a minimum length of 4 (if the input were "1238"), and the regex "\d{1-2}-\d{1-2}-\d{2-4}" has a minimum length of 6 (e.g. "1-10-20"). If the input string is shorter than that minimum length, FindFirstChar will now immediately return false, as it can't possibly match, avoiding spending time evaluating a pattern that couldn't possibly match.

Contributes to https://github.com/dotnet/runtime/issues/1349
cc: @danmosemsft, @eerhardt, @ViktorHofer, @pgovind 

As a silly example:
```C#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using System.Text.RegularExpressions;

public class Program
{
    static void Main(string[] args) => BenchmarkSwitcher.FromAssemblies(new[] { typeof(Program).Assembly }).Run(args);

    [Params(RegexOptions.None, RegexOptions.Compiled)]
    public RegexOptions Options { get; set; }

    private string _input = new string('x', 28);
    private Regex _regex;

    [GlobalSetup]
    public void Setup() => _regex = new Regex(@".*hello there how are you today", Options);

    [Benchmark] public bool Test() => _regex.IsMatch(_input);
}
```

| Method |           Toolchain |  Options |        Mean |      Error |     StdDev |      Median | Ratio |
|------- |-------------------- |--------- |------------:|-----------:|-----------:|------------:|------:|
|   Test | \master\corerun.exe |     None | 7,692.43 ns | 193.846 ns | 456.919 ns | 7,525.43 ns | 1.000 |
|   Test |     \pr\corerun.exe |     None |    24.13 ns |   0.104 ns |   0.098 ns |    24.13 ns | 0.003 |
|        |                     |          |             |            |            |             |       |
|   Test | \master\corerun.exe | Compiled | 2,141.20 ns |  12.438 ns |  11.026 ns | 2,141.36 ns |  1.00 |
|   Test |     \pr\corerun.exe | Compiled |    24.38 ns |   0.124 ns |   0.110 ns |    24.37 ns |  0.01 |