@stephentoub 2020-01-12
(I'm going to tweak this to add it to the interpreter as well.)