@danmosemsft 2020-01-13
BTW, what does RegexFCD mean? First Character Detector?