@stephentoub 2020-01-13
> BTW, what does RegexFCD mean? First Character Detector?

I honestly have no idea about the 'D'. :)  The "FC" does mean "First Characters", at least based on context.  I've gone through and cleaned up a decent amount of such stuff previously, but there's still more to be done. :)